LPS-190119 - move ScimClientUtil REST api module so it is accessible to web
