+++ modules/dxp/apps/scim/scim-rest-api/bnd.bnd
@@ -3,4 +3,5 @@ Bundle-SymbolicName: com.liferay.scim.rest.api
 Bundle-Version: 2.0.0
 Export-Package:\
 	com.liferay.scim.rest.dto.v1_0,\
-	com.liferay.scim.rest.resource.v1_0
+	com.liferay.scim.rest.resource.v1_0,\
+	com.liferay.scim.rest.util
