+++ modules/dxp/apps/scim/scim-rest-api/src/main/resources/com/liferay/scim/rest/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
